Revert "Change expiresIn's type from int to long"
This reverts commit 5a24101acc21b3f552d87f82280d03de0cc03692.

+++ trip-kit-booking/src/main/java/com/skedgo/android/tripkit/booking/ExternalOAuth.java
@@ -15,7 +15,7 @@ public abstract class ExternalOAuth {
   public abstract String authServiceId();
   public abstract String token();
   @Nullable public abstract String refreshToken();
-  public abstract long expiresIn();
+  public abstract int expiresIn();
 
   @Value.Default public String id() {
     return UUID.randomUUID().toString();
